# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Rotina Viva API
  version: 0.1.0
  description: |
    Contrato HTTP entre o frontend Next.js e o worker FastAPI.

    **Fase 0 (monorepo):** rotas documentadas; implementação completa nas fases 1–4.
    Autenticação evolui de login/sessão (PoC) para JWT Supabase (Fase 1).
  contact:
    name: Br Lab
    url: https://github.com/Leonardolabdc/rotina_viva_pro
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8000
    description: FastAPI local (dev)
  - url: https://api.rotinaviva.example
    description: Worker em produção (Railway/Fly — placeholder)

tags:
  - name: health
    description: Disponibilidade e versão
  - name: auth
    description: Sessão e perfil (→ Supabase Auth na Fase 1)
  - name: chat
    description: Assistente IA — plano, SQL, RAG, CrewAI, streaming
  - name: reports
    description: Relatórios sono e alimentação
  - name: transcribe
    description: Voz → texto (Whisper / OpenRouter STT)
  - name: direct-chat
    description: Mensagens família ↔ educador (sem LLM)
  - name: students
    description: Cadastro de alunos (escopo RBAC)

security:
  - bearerAuth: []

paths:
  /health:
    get:
      tags: [health]
      operationId: getHealth
      summary: Health check
      security: []
      responses:
        "200":
          description: API operacional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: Serviço indisponível
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Autenticar utilizador
      security: []
      description: |
        PoC: valida contra `rotina_users.json`. Fase 1: Supabase Auth (email/senha ou magic link).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Sessão criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Terminar sessão
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Sessão invalidada
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: Perfil do utilizador autenticado
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Perfil RBAC
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /chat/sessions:
    post:
      tags: [chat]
      operationId: createChatSession
      summary: Criar sessão de chat
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatSessionRequest"
      responses:
        "201":
          description: Sessão criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /chat/sessions/{sessionId}:
    get:
      tags: [chat]
      operationId: getChatSession
      summary: Obter sessão e histórico
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/sessionId"
      responses:
        "200":
          description: Sessão com mensagens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSession"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /chat/sessions/{sessionId}/messages:
    post:
      tags: [chat]
      operationId: sendChatMessage
      summary: Enviar mensagem (resposta síncrona)
      description: |
        Alternativa não-streaming. Para UX de chat, preferir `POST .../stream`.
        Aplica guardrails de entrada/saída, quota LLM e RBAC antes/depois do motor em `src/`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/sessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageRequest"
      responses:
        "200":
          description: Resposta do assistente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/GuardrailBlocked"
        "429":
          $ref: "#/components/responses/QuotaExceeded"

  /chat/sessions/{sessionId}/messages/stream:
    post:
      tags: [chat]
      operationId: streamChatMessage
      summary: Enviar mensagem (SSE)
      description: |
        Server-Sent Events com tokens da resposta. Eventos: `status`, `token`, `done`, `error`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/sessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageRequest"
      responses:
        "200":
          description: Stream SSE
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Formato: `event: status\ndata: {"phase":"rag"}\n\n`
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/GuardrailBlocked"
        "429":
          $ref: "#/components/responses/QuotaExceeded"

  /reports/sleep-meal:
    get:
      tags: [reports]
      operationId: getSleepMealReport
      summary: Relatório sono e refeições por aluno
      security:
        - bearerAuth: []
      parameters:
        - name: studentName
          in: query
          required: true
          schema:
            type: string
          description: Nome ou parte do nome (gestão/educador). Família usa o aluno vinculado.
        - name: studentId
          in: query
          schema:
            type: integer
          description: Opcional — id_aluno quando conhecido
      responses:
        "200":
          description: Dados tabulares + metadados para gráficos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepMealReport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /transcribe:
    post:
      tags: [transcribe]
      operationId: transcribeAudio
      summary: Transcrever áudio para texto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                language:
                  type: string
                  default: pt
      responses:
        "200":
          description: Texto transcrito
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscribeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "413":
          description: Ficheiro demasiado grande
        "502":
          description: Serviço STT indisponível

  /direct-chat/students:
    get:
      tags: [direct-chat]
      operationId: listDirectChatStudents
      summary: Alunos disponíveis para chat directo
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista escopada por perfil
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /direct-chat/students/{studentId}/messages:
    get:
      tags: [direct-chat]
      operationId: listDirectChatMessages
      summary: Histórico família ↔ educador
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/studentId"
      responses:
        "200":
          description: Mensagens ordenadas por tempo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DirectChatMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [direct-chat]
      operationId: sendDirectChatMessage
      summary: Enviar mensagem directa
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/studentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendDirectChatMessageRequest"
      responses:
        "201":
          description: Mensagem registada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectChatMessage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /students:
    get:
      tags: [students]
      operationId: listStudents
      summary: Listar alunos (escopo RBAC)
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Filtro por nome
      responses:
        "200":
          description: Alunos visíveis ao perfil
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Fase 0: token de sessão opaco. Fase 1+: JWT Supabase (`Authorization: Bearer`).

  parameters:
    sessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    studentId:
      name: studentId
      in: path
      required: true
      schema:
        type: integer

  responses:
    Unauthorized:
      description: Não autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: RBAC — perfil sem permissão
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    GuardrailBlocked:
      description: Guardrail de entrada bloqueou a mensagem
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/GuardrailVerdict"
    QuotaExceeded:
      description: Quota diária de mensagens LLM excedida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    HealthResponse:
      type: object
      required: [status, version, phase]
      properties:
        status:
          type: string
          enum: [ok, degraded]
        version:
          type: string
          example: "0.1.0"
        phase:
          type: string
          description: Fase do roadmap de migração
          example: "0-monorepo"

    RotinaRole:
      type: string
      enum: [gestao, educador, familia]

    DataSourceMode:
      type: string
      enum: [auto, structured, documents]
      description: Equivalente a `data_source_mode` no Streamlit

    UserProfile:
      type: object
      required: [username, role, displayName]
      properties:
        username:
          type: string
        role:
          $ref: "#/components/schemas/RotinaRole"
        displayName:
          type: string
        studentId:
          type: integer
          description: Apenas perfil família — id_aluno vinculado
        allowMutations:
          type: boolean
          description: Escrita em cadastro/diário (gestão; educador conforme RBAC)

    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password

    AuthSession:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
        expiresAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/UserProfile"

    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    ChatSession:
      type: object
      required: [id, messages]
      properties:
        id:
          type: string
          format: uuid
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        dataSourceMode:
          $ref: "#/components/schemas/DataSourceMode"
        crewAiEnabled:
          type: boolean
        predictiveMlEnabled:
          type: boolean

    CreateChatSessionRequest:
      type: object
      properties:
        dataSourceMode:
          $ref: "#/components/schemas/DataSourceMode"
        crewAiEnabled:
          type: boolean
          default: false
        predictiveMlEnabled:
          type: boolean
          default: false

    SendChatMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        dataSourceMode:
          $ref: "#/components/schemas/DataSourceMode"
        crewAiEnabled:
          type: boolean
        predictiveMlEnabled:
          type: boolean
        confirmMutation:
          type: boolean
          description: Gestão confirma DELETE/UPDATE sensível (equivalente UI Streamlit)

    RagChunk:
      type: object
      properties:
        source:
          type: string
        chunk:
          type: string
        distance:
          type: number
        text:
          type: string

    GuardrailVerdict:
      type: object
      required: [allowed, stage]
      properties:
        allowed:
          type: boolean
        stage:
          type: string
          enum: [input, output]
        reason:
          type: string
        scanner:
          type: string
        redactedContent:
          type: string

    ChatMessageResponse:
      type: object
      required: [message]
      properties:
        message:
          $ref: "#/components/schemas/ChatMessage"
        ragChunks:
          type: array
          items:
            $ref: "#/components/schemas/RagChunk"
        guardrail:
          $ref: "#/components/schemas/GuardrailVerdict"
        processingStatus:
          type: string
          description: Linha de status (SQL/RAG/planeador) para debug na UI

    SleepMealReport:
      type: object
      required: [studentLabel, summaryMarkdown, rows]
      properties:
        studentLabel:
          type: string
        summaryMarkdown:
          type: string
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
        sleepReference:
          type: object
          additionalProperties: true
        chartSpec:
          type: object
          description: Especificação Vega-Lite / Altair serializada (opcional)

    TranscribeResponse:
      type: object
      required: [text]
      properties:
        text:
          type: string
        language:
          type: string

    StudentSummary:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        className:
          type: string

    DirectChatMessage:
      type: object
      required: [id, studentId, senderRole, content, createdAt]
      properties:
        id:
          type: string
          format: uuid
        studentId:
          type: integer
        senderRole:
          $ref: "#/components/schemas/RotinaRole"
        content:
          type: string
        createdAt:
          type: string
          format: date-time

    SendDirectChatMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1

    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true